Fixed: Only one machine image version per vendor visible in dashboard
